requirements.txt
discord.py
mypy
python_dotenv
pillow
httpx
pymongo
dnspython
graphene
FastAPI
uvicorn
pytest
requests
PyJWT
